+++ docs-website/config/navigation.js
@@ -63,6 +63,19 @@ const navigation = [
 		],
 	},
 	{
+		title: 'Pre 1.0 Changes',
+		links: [
+			{
+				title: 'Overview',
+				href: '/docs/pre-1-0-changes',
+			},
+			{
+				title: 'Simplifying the WunderGraph URL Structure',
+				href: '/docs/pre-1-0-changes/simplifying-the-wundergraph-url-structure',
+			},
+		],
+	},
+	{
 		title: 'Architecture',
 		links: [
 			{
@@ -146,6 +159,15 @@ const navigation = [
 		],
 	},
 	{
+		title: 'Cloud',
+		links: [
+			{
+				title: 'Deploy to WunderGraph Cloud',
+				href: '/docs/cloud/deployments',
+			},
+		],
+	},
+	{
 		title: 'Deployment',
 		links: [
 			{
@@ -174,8 +196,8 @@ const navigation = [
 				href: '/docs/examples/nextjs',
 			},
 			{
-				title: 'NextJS + SWR',
-				href: '/docs/examples/nextjs-swr',
+				title: 'Vite + SWR',
+				href: '/docs/examples/vite-swr',
 			},
 			{
 				title: 'Hooks',
@@ -526,6 +548,10 @@ const navigation = [
 			{
 				title: 'Overview',
 				href: '/docs/components-of-wundergraph',
+			},
+			{
+				title: 'create-wundergraph-app',
+				href: '/docs/components-of-wundergraph/create-wundergraph-app',
 			},
 			{
 				title: 'wunderctl',