+++ examples/migrate-from-apollo/pages/_app.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client';
 
 const client = new ApolloClient({
-	uri: 'http://localhost:9991/app/main/graphql',
+	uri: 'http://localhost:9991/graphql',
 	cache: new InMemoryCache(),
 });