+++ examples/nextjs-postgres-user-filters/.wundergraph/wundergraph.config.ts
@@ -1,11 +1,4 @@
-import {
-	Application,
-	authProviders,
-	configureWunderGraphApplication,
-	cors,
-	introspect,
-	templates,
-} from '@wundergraph/sdk';
+import { authProviders, configureWunderGraphApplication, cors, introspect, templates } from '@wundergraph/sdk';
 import { NextJsTemplate } from '@wundergraph/nextjs/dist/template';
 import server from './wundergraph.server';
 import operations from './wundergraph.operations';
@@ -15,27 +8,17 @@ const db = introspect.postgresql({
 	databaseURL: 'postgresql://admin:admin@localhost:54322/example?schema=public',
 });
 
-const myApplication = new Application({
-	name: 'app',
-	apis: [db],
-});
-
 // configureWunderGraph emits the configuration
 configureWunderGraphApplication({
-	application: myApplication,
+	apis: [db],
 	server,
 	operations,
 	codeGenerators: [
 		{
-			templates: [
-				...templates.typescript.all,
-				templates.typescript.operations,
-				templates.typescript.linkBuilder,
-				templates.typescript.client,
-			],
+			templates: [...templates.typescript.all, templates.typescript.operations, templates.typescript.linkBuilder],
 		},
 		{
-			templates: [new NextJsTemplate()],
+			templates: [templates.typescript.client, new NextJsTemplate()],
 			path: '../components/generated',
 		},
 	],