+++ packages/protobuf/wgpb/wundernode_config.ts
@@ -776,6 +776,7 @@ export interface DataSourceCustomGraphQL {
   federation: GraphQLFederationConfiguration | undefined;
   upstreamSchema: string;
   hooksConfiguration: GraphQLDataSourceHooksConfiguration | undefined;
+  customScalarTypeFields: SingleTypeField[];
 }
 
 export interface DataSourceCustomDatabase {
@@ -908,9 +909,7 @@ export interface ArgumentConfiguration {
 export interface WunderGraphConfiguration {
   api: UserDefinedApi | undefined;
   apiId: string;
-  deploymentName: string;
   environmentIds: string[];
-  apiName: string;
   dangerouslyEnableGraphQLEndpoint: boolean;
 }
 
@@ -2348,6 +2347,7 @@ function createBaseDataSourceCustomGraphQL(): DataSourceCustomGraphQL {
     federation: undefined,
     upstreamSchema: "",
     hooksConfiguration: undefined,
+    customScalarTypeFields: [],
   };
 }
 
@@ -2363,6 +2363,9 @@ export const DataSourceCustomGraphQL = {
       hooksConfiguration: isSet(object.hooksConfiguration)
         ? GraphQLDataSourceHooksConfiguration.fromJSON(object.hooksConfiguration)
         : undefined,
+      customScalarTypeFields: Array.isArray(object?.customScalarTypeFields)
+        ? object.customScalarTypeFields.map((e: any) => SingleTypeField.fromJSON(e))
+        : [],
     };
   },
 
@@ -2378,6 +2381,11 @@ export const DataSourceCustomGraphQL = {
     message.hooksConfiguration !== undefined && (obj.hooksConfiguration = message.hooksConfiguration
       ? GraphQLDataSourceHooksConfiguration.toJSON(message.hooksConfiguration)
       : undefined);
+    if (message.customScalarTypeFields) {
+      obj.customScalarTypeFields = message.customScalarTypeFields.map((e) => e ? SingleTypeField.toJSON(e) : undefined);
+    } else {
+      obj.customScalarTypeFields = [];
+    }
     return obj;
   },
 
@@ -2396,6 +2404,7 @@ export const DataSourceCustomGraphQL = {
     message.hooksConfiguration = (object.hooksConfiguration !== undefined && object.hooksConfiguration !== null)
       ? GraphQLDataSourceHooksConfiguration.fromPartial(object.hooksConfiguration)
       : undefined;
+    message.customScalarTypeFields = object.customScalarTypeFields?.map((e) => SingleTypeField.fromPartial(e)) || [];
     return message;
   },
 };
@@ -3118,14 +3127,7 @@ export const ArgumentConfiguration = {
 };
 
 function createBaseWunderGraphConfiguration(): WunderGraphConfiguration {
-  return {
-    api: undefined,
-    apiId: "",
-    deploymentName: "",
-    environmentIds: [],
-    apiName: "",
-    dangerouslyEnableGraphQLEndpoint: false,
-  };
+  return { api: undefined, apiId: "", environmentIds: [], dangerouslyEnableGraphQLEndpoint: false };
 }
 
 export const WunderGraphConfiguration = {
@@ -3133,9 +3135,7 @@ export const WunderGraphConfiguration = {
     return {
       api: isSet(object.api) ? UserDefinedApi.fromJSON(object.api) : undefined,
       apiId: isSet(object.apiId) ? String(object.apiId) : "",
-      deploymentName: isSet(object.deploymentName) ? String(object.deploymentName) : "",
       environmentIds: Array.isArray(object?.environmentIds) ? object.environmentIds.map((e: any) => String(e)) : [],
-      apiName: isSet(object.apiName) ? String(object.apiName) : "",
       dangerouslyEnableGraphQLEndpoint: isSet(object.dangerouslyEnableGraphQLEndpoint)
         ? Boolean(object.dangerouslyEnableGraphQLEndpoint)
         : false,
@@ -3146,13 +3146,11 @@ export const WunderGraphConfiguration = {
     const obj: any = {};
     message.api !== undefined && (obj.api = message.api ? UserDefinedApi.toJSON(message.api) : undefined);
     message.apiId !== undefined && (obj.apiId = message.apiId);
-    message.deploymentName !== undefined && (obj.deploymentName = message.deploymentName);
     if (message.environmentIds) {
       obj.environmentIds = message.environmentIds.map((e) => e);
     } else {
       obj.environmentIds = [];
     }
-    message.apiName !== undefined && (obj.apiName = message.apiName);
     message.dangerouslyEnableGraphQLEndpoint !== undefined &&
       (obj.dangerouslyEnableGraphQLEndpoint = message.dangerouslyEnableGraphQLEndpoint);
     return obj;
@@ -3164,9 +3162,7 @@ export const WunderGraphConfiguration = {
       ? UserDefinedApi.fromPartial(object.api)
       : undefined;
     message.apiId = object.apiId ?? "";
-    message.deploymentName = object.deploymentName ?? "";
     message.environmentIds = object.environmentIds?.map((e) => e) || [];
-    message.apiName = object.apiName ?? "";
     message.dangerouslyEnableGraphQLEndpoint = object.dangerouslyEnableGraphQLEndpoint ?? false;
     return message;
   },